moved CalculationParameters and SCFProtocol into separate files
